refactoring: rearrange code in files
