+++ cursive/examples/mines/board/model.rs
